layout/style/test/test_bug1451199-2.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1451199
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1451199</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1451199 **/
  SimpleTest.waitForExplicitFinish();

  addLoadEvent(async function() {
    // We have to be careful to not check l.matches until the very end
    // of the test.
    const l = frames[0].matchMedia("(orientation: portrait)");
    const iframe = document.querySelector("iframe");
    iframe.width = "50";

    await new Promise(r => requestAnimationFrame(() => requestAnimationFrame(r)));

    l.addEventListener("change", function() {
      is(l.matches, false,
         "Should not match portrait by the time we get notified");
      SimpleTest.finish();
    }, { once: true });
    iframe.width = "200";
  });
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1451199">Mozilla Bug 1451199</a>
<p id="display"><iframe height="100" width="200"></iframe></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
